TPC Workflow: Move TPC readers into separate TPCReaderWorkflow library
