+++ Detectors/GlobalTrackingWorkflow/helpers/CMakeLists.txt
@@ -20,6 +20,7 @@ o2_add_library(GlobalTrackingWorkflowHelpers
                                       O2::FT0Workflow
                                       O2::ITSMFTWorkflow
                                       O2::TRDWorkflowIO
+                                      O2::TPCReaderWorkflow
                                       O2::SimulationDataFormat)
 
 o2_add_executable(track-cluster-reader
